+++ lib/widgets/amount.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:money/helpers/color_helper.dart';
 import 'package:money/models/money_objects/currencies/currency.dart';
 
 /// Formatted text using the supplied currency code and optional the currency/country flag
@@ -9,17 +10,31 @@ class Amount extends StatelessWidget {
   /// USD | CAD | GBP
   final String iso4217;
 
+  final bool showCurrency;
+  final bool autoColor;
+
   /// Constructor
-  const Amount(this.value, this.iso4217, {super.key});
+  const Amount(this.value, this.iso4217, {super.key, this.showCurrency = true, this.autoColor = false});
 
   @override
   Widget build(final BuildContext context) {
-    return Row(
-      children: [
-        Text(Currency.getAmountAsStringUsingCurrency(value)),
-        const SizedBox(width: 10),
-        Currency.buildCurrencyWidget(iso4217),
-      ],
+    if (showCurrency) {
+      return Row(
+        children: [
+          amountAsText(),
+          const SizedBox(width: 10),
+          Currency.buildCurrencyWidget(iso4217),
+        ],
+      );
+    } else {
+      return amountAsText();
+    }
+  }
+
+  Widget amountAsText() {
+    return Text(
+      Currency.getAmountAsStringUsingCurrency(value),
+      style: TextStyle(fontFamily: 'RobotoMono', color: autoColor ? colorBasedOnValue(value) : null),
     );
   }
 }